pyproject.toml
[project]
name = "openclaw-skillify"
version = "0.1.0"
description = "Turn any API surface into OpenClaw/nanobot skills"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["openclaw", "nanobot", "ai", "skills", "api", "openapi", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "litellm>=1.50",
    "typer>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/openclaw-skillify"
Repository = "https://github.com/YOUR_USERNAME/openclaw-skillify"
Issues = "https://github.com/YOUR_USERNAME/openclaw-skillify/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[project.scripts]
skillify = "skillify.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skillify"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
